Converted courses page to TS
